pyproject.toml
[project]
name = "sp-14-blue-chess-ai"
version = "0.1.0"
description = "Chess AI engine. Play against AI or another player."
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "torch (>=2.6.0,<3.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "pyinstaller (>=6.13.0,<7.0.0)",
    "pywebview (>=5.4,<6.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false